requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Sentiment analysis libraries
textblob>=0.17.1
vaderSentiment>=3.3.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTML generation
jinja2>=3.1.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and utilities
colorama>=0.4.6
tqdm>=4.65.0 